Simple Word Count of the EKU Student Handbook
Word Count using MapReduce framework in Hadoop, in Manjaro Linux. This part of the code reads the txt file that was processed using pdf2txt.py and preprocessed for zipf's law analysis, remove common words (stopwords), map and reduce, this part of the project was the only part that successfully ran on hdfs.